+++ R/conclass.R
@@ -122,8 +122,9 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
     #' @param k balancing.factor.per.sample (default=NULL)
     #' @return joint graph to be used for downstream analysis
     buildGraph=function(k=15, k.self=10, k.self.weight=0.1, alignment.strength=NULL, space='PCA', matching.method='mNN', metric='angular', k1=k, data.type='counts', l2.sigma=1e5, var.scale=TRUE, ncomps=40,
-                        n.odgenes=2000, neighborhood.average=FALSE, neighborhood.average.k=10, matching.mask=NULL, exclude.samples=NULL, common.centering=TRUE, verbose=TRUE,
+                        n.odgenes=2000, matching.mask=NULL, exclude.samples=NULL, common.centering=TRUE, verbose=TRUE,
                         base.groups=NULL, append.global.axes=TRUE, append.decoys=TRUE, decoy.threshold=1, n.decoys=k*2, score.component.variance=FALSE,
+                        snn= FALSE, snn.quantile=0.9,min.snn.jaccard=0,min.snn.weight=0, snn.k=k.self,
                         balance.edge.weights=FALSE, balancing.factor.per.cell=NULL, same.factor.downweight=1.0, k.same.factor=k, balancing.factor.per.sample=NULL) {
       supported.spaces <- c("CPCA","JNMF","genes","PCA","PMA","CCA")
       if(!space %in% supported.spaces) {
@@ -137,6 +138,17 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       if(!metric %in% supported.metrics) {
         stop(paste0("only the following distance metrics are currently supported: ['",paste(supported.metrics,collapse="' '"),"']"))
       }
+
+      if(!is.null(snn.quantile) && !is.na(snn.quantile)) {
+        if(length(snn.quantile)==1)  {
+          snn.quantile <- c(1-snn.quantile,snn.quantile)
+        } 
+        snn.quantile <- sort(snn.quantile,decreasing=F)
+        if(snn.quantile[1]<0 | snn.quantile[2]>1) {
+          stop("snn.quantile must be one or two numbers in the [0,1] range")
+        }
+      }
+      
       if (!is.null(alignment.strength)) {
         alignment.strength %<>% max(0) %>% min(1)
         k1 <- sapply(self$samples, function(sample) ncol(getCountMatrix(sample))) %>% max() %>%
@@ -148,8 +160,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       if(k1<k) { stop("k1 must be >= k") }
       # calculate or update pairwise alignments
       sn.pairs <- private$updatePairs(
-        space=space, ncomps=ncomps, n.odgenes=n.odgenes, verbose=verbose, var.scale=var.scale, neighborhood.average=neighborhood.average,
-        neighborhood.average.k=10, matching.mask=matching.mask, exclude.samples=exclude.samples, score.component.variance=score.component.variance
+        space=space, ncomps=ncomps, n.odgenes=n.odgenes, verbose=verbose, var.scale=var.scale, matching.mask=matching.mask, exclude.samples=exclude.samples, score.component.variance=score.component.variance
       )
       if(ncol(sn.pairs)<1) { stop("insufficient number of comparable pairs") }
       if(!is.null(base.groups)) {
@@ -160,9 +171,17 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
         samf <- setNames(rep(names(samf),unlist(lapply(samf,length))),unlist(samf))
         if(append.global.axes) {
           cms.clust <- self$getClusterCountMatrices(groups=base.groups,common.genes=FALSE)
-          global.proj <- projectSamplesOnGlobalAxes(self$samples, cms.clust, data.type, neighborhood.average, verbose, self$n.cores)
+          global.proj <- projectSamplesOnGlobalAxes(self$samples, cms.clust, data.type, verbose, self$n.cores)
         }
       }
+
+
+      if(snn) {
+        local.neighbors <- getLocalNeighbors(self$samples[! names(self$samples) %in% exclude.samples], snn.k, k.self.weight, metric, l2.sigma=l2.sigma, verbose, self$n.cores)
+      } else {
+        local.neighbors <- NULL
+      }
+
 
       # determine inter-sample mapping
       if(verbose) cat('inter-sample links using ',matching.method,' ');
@@ -206,7 +225,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
 
           mnn <- getPcaBasedNeighborMatrix(self$samples[sn.pairs[,j]], od.genes=od.genes, rot=rot, data.type=data.type,
                                            k=k.cur, k1=k1, matching.method=matching.method, metric=metric, l2.sigma=l2.sigma, cor.base=cor.base,
-                                           var.scale=var.scale, neighborhood.average=neighborhood.average, common.centering=common.centering,
+                                           var.scale=var.scale, common.centering=common.centering,
                                            base.groups=base.groups, append.decoys=append.decoys, samples=self$samples, samf=samf, decoy.threshold=decoy.threshold,
                                            n.decoys=n.decoys, append.global.axes=append.global.axes, global.proj=global.proj)
         } else if (space=='genes') { ## Overdispersed Gene space
@@ -219,6 +238,37 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
           stop("Unknown space: ", space)
         }
 
+        if(snn) { # optionally, perform shared neighbor weighting, a la SeuratV3, scran
+          
+          m1 <- cbind(mnn,t(local.neighbors[[sn.pairs[1,j] ]]))
+          m2 <- rbind(local.neighbors[[sn.pairs[2,j] ]], mnn)
+
+          mnn1 <- mnn; mnn1@x <- rep(1,length(mnn1@x))
+          m1@x <- rep(1,length(m1@x)); m2@x <- rep(1,length(m2@x))
+
+          x <- ((m1 %*% m2) * mnn1) / pmax(outer(rowSums(m1),colSums(m2),FUN=pmin),1);
+          
+          # scale by Jaccard coefficient
+
+          if(min.snn.jaccard>0) {
+            x@x[x@x<min.snn.jaccard] <- 0;
+          }
+          
+          if(!is.null(snn.quantile) && !is.na(snn.quantile)) {
+            xq <- quantile(x@x,p=c(snn.quantile[1],snn.quantile[2]))
+            x@x <- pmax(0,pmin(1,(x@x-xq[1])/pmax(1,diff(xq))))
+          }
+          
+          x <- drop0(x)
+          if(min.snn.weight>0) {
+            mnn <- as(drop0(mnn*min.snn.weight + mnn*x),'dgTMatrix')
+          } else {
+            mnn <- as(drop0(mnn*x),'dgTMatrix')
+          }
+
+        }
+
+        
         if(verbose) cat(".")
         return(data.frame('mA.lab'=rownames(mnn)[mnn@i+1],'mB.lab'=colnames(mnn)[mnn@j+1],'w'=mnn@x, stringsAsFactors=FALSE))
       },n.cores=self$n.cores,mc.preschedule=TRUE)
@@ -227,17 +277,21 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       ## Merge the results into a edge table
       el <- do.call(rbind,mnnres)
       el$type <- 1; # encode connection type 1- intersample, 0- intrasample
-      # append some local edges
+      # append local edges
       if(k.self>0) {
-        if(verbose) cat('local pairs ')
-        x <- getLocalEdges(self$samples[! names(self$samples) %in% exclude.samples], k.self, k.self.weight, metric, l2.sigma=l2.sigma, verbose, self$n.cores)
-        el <- rbind(el,x)
+        if(is.null(local.neighbors) || snn.k!=k.self) { # recalculate local neighbors
+          local.neighbors <- getLocalNeighbors(self$samples[! names(self$samples) %in% exclude.samples], k.self, k.self.weight, metric, l2.sigma=l2.sigma, verbose, self$n.cores)
+        }
+        el <- rbind(el,getLocalEdges(local.neighbors))
       }
       if(verbose) cat('building graph .')
+      el <- el[el[,3]>0,];
       g  <- graph_from_edgelist(as.matrix(el[,c(1,2)]), directed =FALSE)
       E(g)$weight <- el[,3]
       E(g)$type <- el[,4]
+      
       if(verbose) cat('.')
+
       # collapse duplicate edges
       g <- simplify(g, edge.attr.comb=list(weight="sum", type = "first"))
       if(verbose) cat('done\n')
@@ -269,6 +323,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       return(invisible(g))
     },
 
+
     #' @description  Calculates differential genes. Estimates base mean, z-score, p-values, specificity, precision, expressionFraction, AUC (if append.auc=TRUE)
     #'
     #' @param z.threshold (default=3.0)
@@ -280,6 +335,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
     #' @return list of DE results
     getDifferentialGenes=function(clustering=NULL, groups=NULL, z.threshold=3.0, upregulated.only=FALSE, verbose=TRUE, plot=FALSE, n.genes.to.show=10, inner.clustering=FALSE,
                                   append.specificity.metrics=TRUE, append.auc=FALSE, n.cores=self$n.cores) {
+
       groups <- parseCellGroups(self, clustering, groups)
 
       groups %<>% as.factor() %>% droplevels()
@@ -301,7 +357,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
 
         de.genes %<>% lapply(function(x) if ((length(x) > 0) && (nrow(x) > 0)) subset(x, complete.cases(x)) else x)
         de.genes %<>% names() %>% setNames(., .) %>%
-          plapply(function(n) appendSpecificityMetricsToDE(de.genes[[n]], groups.clean, n, p2.counts=cm.merged, append.auc=append.auc), verbose=verbose, n.cores=n.cores)
+          sccore::plapply(function(n) appendSpecificityMetricsToDE(de.genes[[n]], groups.clean, n, p2.counts=cm.merged, append.auc=append.auc), progress=verbose, n.cores=n.cores)
       }
 
       if (verbose) cat("All done!\n")
@@ -366,7 +422,9 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
             i2 <- which(p1==p1[i1[[1]]])
             length(intersect(i1,i2))/length(unique(c(i1,i2)))
           })
+
         },n.cores=self$n.cores, mc.preschedule=TRUE))
+
 
         # Adjusted rand index
         if(verbose) cat("adjusted Rand ... ")
@@ -400,14 +458,16 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
           # assess stability for that hierarchy (to visualize internal node stability)
           # for the original clustering and every subsample clustering,
           if(verbose) cat("upper clustering ... ")
-          cgraph <- getClusterGraph(self$graph,cls.groups,plot=F,normalize=F)
+          cgraph <- getClusterGraph(self$graph,cls.groups,plot=FALSE,normalize=FALSE)
           chwt <- walktrap.community(cgraph,steps=9)
           clm <- igraph:::complete.dend(chwt,FALSE)
 
           if(verbose) cat("clusterTree Jaccard ... ")
           jc.hstats <- do.call(rbind, papply(sr,function(st1) {
             mf <- membership(st1); mf <- as.factor(setNames(as.character(mf),names(mf)))
+
             st1g <- getClusterGraph(self$graph,mf, plot=FALSE, normalize=TRUE)
+
             st1w <- walktrap.community(st1g, steps=8)
 
             #merges <- st1w$merge; leaf.factor <- mf; clusters <- cls.groups
@@ -491,7 +551,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
         if (!requireNamespace("uwot", quietly=TRUE))
           stop("You need to install package 'uwot' to be able to use UMAP embedding.")
 
-        self$embedding <- embedGraphUmap(self$graph, verbose=verbose, return.all=F, n.cores=n.cores, target.dims=target.dims, ...)
+        self$embedding <- embedGraphUmap(self$graph, verbose=verbose, return.all=FALSE, n.cores=n.cores, target.dims=target.dims, ...)
       }
 
       return(invisible(self$embedding))
@@ -573,7 +633,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
         }
         xy <- dendextend::get_nodes_xy(d)
         to <- t.dfirst(hc$merge)
-        plot(d,las=2,axes=F)
+        plot(d,las=2,axes=FALSE)
         # flat on the left
         #x <- apply(st$flat$jc,2,median)
         #text(xy,labels=round(x[to],2),col='blue',adj=c(-0.1,-1.24),cex=0.8)
@@ -597,7 +657,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
 
       emb <- self$embedding;
       if(!is.null(subset)) {
-        emb <- emb[rownames(emb) %in% subset,,drop=F]
+        emb <- emb[rownames(emb) %in% subset,,drop=FALSE]
       }
 
       if (!is.null(gene)) {
@@ -628,7 +688,8 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
     #' @param max.iters maximal number of diffusion iterations. Default: 15.
     #' @param tol tolerance after which the diffusion stops. Default: 5e-3.
     #' @param name name to save the correction. Default: diffusion.
-    #' @param count.matrix alternative gene count matrix to correct. Default: joint count matrix for all datasets.
+    #' @param verbose verbose mode. Default: TRUE.
+    #' @param count.matrix alternative gene count matrix to correct (rows: genes, columns: cells; has to be dense matrix). Default: joint count matrix for all datasets.
     correctGenes=function(genes=NULL, n.od.genes=500, fading=10.0, fading.const=0.5, max.iters=15, tol=5e-3, name='diffusion', verbose=TRUE, count.matrix=NULL, normalize=TRUE) {
       edges <- igraph::as_edgelist(self$graph)
       edge.weights <- igraph::edge.attributes(self$graph)$weight
@@ -643,6 +704,13 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
         count.matrix <- Reduce(rbind, lapply(cms, function(x) x[, genes])) %>% as.matrix()
       } else {
         count.matrix <- t(count.matrix)
+      }
+      vn <- V(self$graph)$name;
+      if(!all(rownames(count.matrix)==vn)) { # subset to a common set of genes
+        if(!all(vn %in% rownames(count.matrix))) {
+          stop("count.matrix does not provide values for all the vertices in the alignment graph!")
+        }
+        count.matrix <- count.matrix[vn,]
       }
 
       cm <- smoothMatrixOnGraph(edges, edge.weights, count.matrix, max_n_iters=max.iters, diffusion_fading=fading,
@@ -689,7 +757,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
         m <- conos:::getRawCountMatrix(s,trans=TRUE); # rows are cells
         cl <- factor(groups[match(rownames(m),names(groups))],levels=levels(groups));
         tc <- colSumByFactor(m,cl);
-        if(omit.na.cells) { tc <- tc[-1,,drop=F] }
+        if(omit.na.cells) { tc <- tc[-1,,drop=FALSE] }
         t(tc);
       })
       # bring to a common gene space
@@ -738,25 +806,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       return(main.theme + theme)
     },
 
-    updatePairs=function(space='PCA', data.type='counts', ncomps=50, n.odgenes=1e3, var.scale=TRUE, neighborhood.average=FALSE, neighborhood.average.k=10, matching.mask=NULL, exclude.samples=NULL, score.component.variance=FALSE, verbose=FALSE) {
-      if(neighborhood.average) {
-        # pre-calculate averaging matrices for each sample
-        if(verbose)  cat("calculating local averaging neighborhoods ")
-        for (n in names(self$samples)) {
-          r <- self$samples[[n]]
-          if(!is.null(edgeMat(r)$mat) && edgeMat(r)$k != neighborhood.average.k)
-            next
-
-          xk <- n2Knn(getPca(r)[getCellNames(r),],neighborhood.average.k,self$n.cores,FALSE)
-          xk@x <- pmax(1-xk@x,0);
-          diag(xk) <- 1;
-          xk <- t(t(xk)/colSums(xk))
-          colnames(xk) <- rownames(xk) <- getCellNames(r)
-          edgeMat(self$samples[[n]]) <- list(mat=xk, k=neighborhood.average.k)
-          if(verbose) cat(".")
-        }
-        if(verbose) cat(" done\n")
-      }
+    updatePairs=function(space='PCA', data.type='counts', ncomps=50, n.odgenes=1e3, var.scale=TRUE, matching.mask=NULL, exclude.samples=NULL, score.component.variance=FALSE, verbose=FALSE) {
 
       # make a list of all pairs
       sample.names <- names(self$samples);
@@ -788,28 +838,28 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       nm <- match(apply(sn.pairs,2,paste,collapse='.vs.'),names(self$pairs[[space]]));
       mi[which(!is.na(nm))] <- na.omit(nm);
       # try reverse match as well
-      nm <- match(apply(sn.pairs[c(2,1),,drop=F],2,paste,collapse='.vs.'),names(self$pairs[[space]]));
+      nm <- match(apply(sn.pairs[c(2,1),,drop=FALSE],2,paste,collapse='.vs.'),names(self$pairs[[space]]));
       mi[which(!is.na(nm))] <- na.omit(nm);
       if(verbose) cat('found',sum(!is.na(mi)),'out of',length(mi),'cached',space,' space pairs ... ')
       if(any(is.na(mi))) { # some pairs are missing
         if(verbose) cat('running',sum(is.na(mi)),'additional',space,' space pairs ')
         xl2 <- papply(which(is.na(mi)), function(i) {
           if(space=='CPCA') {
-            xcp <- quickCPCA(self$samples[sn.pairs[,i]],data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale,neighborhood.average=neighborhood.average, score.component.variance=score.component.variance)
+            xcp <- quickCPCA(self$samples[sn.pairs[,i]],data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale, score.component.variance=score.component.variance)
           } else if(space=='JNMF') {
-            xcp <- quickJNMF(self$samples[sn.pairs[,i]],data.type=data.type,n.comps=ncomps,n.odgenes=n.odgenes,var.scale=var.scale,verbose=FALSE,max.iter=3e3,neighborhood.average=neighborhood.average)
+            xcp <- quickJNMF(self$samples[sn.pairs[,i]],data.type=data.type,n.comps=ncomps,n.odgenes=n.odgenes,var.scale=var.scale,verbose=FALSE,max.iter=3e3)
           } else if (space == 'genes') {
-            xcp <- quickNULL(p2.objs = self$samples[sn.pairs[,i]], data.type=data.type, n.odgenes=n.odgenes, var.scale = var.scale, verbose = FALSE, neighborhood.average=neighborhood.average);
+            xcp <- quickNULL(p2.objs = self$samples[sn.pairs[,i]], data.type=data.type, n.odgenes=n.odgenes, var.scale = var.scale, verbose = FALSE)
           } else if (space == 'PCA') {
-            xcp <- quickPlainPCA(self$samples[sn.pairs[,i]], data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale,neighborhood.average=neighborhood.average, score.component.variance=score.component.variance)
+            xcp <- quickPlainPCA(self$samples[sn.pairs[,i]], data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale, score.component.variance=score.component.variance)
           } else if (space == 'CCA' || space=='PMA') {
-            xcp <- quickCCA(self$samples[sn.pairs[,i]],data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale,neighborhood.average=neighborhood.average, score.component.variance=score.component.variance,PMA=(space=='PMA'))
+            xcp <- quickCCA(self$samples[sn.pairs[,i]],data.type=data.type,ncomps=ncomps,n.odgenes=n.odgenes,verbose=FALSE,var.scale=var.scale, score.component.variance=score.component.variance,PMA=(space=='PMA'))
           }
           if(verbose) cat('.')
           xcp
         },n.cores=self$n.cores,mc.preschedule=(space=='PCA'));
 
-        names(xl2) <- apply(sn.pairs[,which(is.na(mi)),drop=F],2,paste,collapse='.vs.');
+        names(xl2) <- apply(sn.pairs[,which(is.na(mi)),drop=FALSE],2,paste,collapse='.vs.');
         xl2 <- xl2[!unlist(lapply(xl2,is.null))]
         self$pairs[[space]] <- c(self$pairs[[space]],xl2);
       }
@@ -818,7 +868,7 @@ Conos <- R6::R6Class("Conos", lock_objects=FALSE,
       mi <- rep(NA,ncol(sn.pairs));
       nm <- match(apply(sn.pairs,2,paste,collapse='.vs.'),names(self$pairs[[space]]));
       mi[which(!is.na(nm))] <- na.omit(nm);
-      nm <- match(apply(sn.pairs[c(2,1),,drop=F],2,paste,collapse='.vs.'),names(self$pairs[[space]]));
+      nm <- match(apply(sn.pairs[c(2,1),,drop=FALSE],2,paste,collapse='.vs.'),names(self$pairs[[space]]));
       mi[which(!is.na(nm))] <- na.omit(nm);
       if(any(is.na(mi))) {
         warning("unable to get complete set of pair comparison results")