+++ R/plot.R
@@ -1,4 +1,7 @@
 #' @importFrom dplyr %>%
+#' @importFrom ComplexHeatmap ht_opt
+#' @importFrom ComplexHeatmap Heatmap
+#' @importFrom ComplexHeatmap HeatmapAnnotation
 NULL
 
 #' @export
@@ -34,8 +37,9 @@ getClusteringGroups <- function(clusters, clustering) {
 ##' @param panel.size vector with two numbers, which specified (width, height) of the panel in inches. Ignored if raster == FALSE.
 ##' @param adjust.func function to adjust plots before combining them to single panel. Can be used, for example, to provide color pallette of guides of the plots.
 ##' @param subset a subset of cells to show (vector of cell names)
+##' @param return.plotlist return a list of ggplot objects instead of a combined plot (default:FALSE)
 ##' @return ggplot2 object with the panel of plots
-plotEmbeddings <- function(embeddings, groups=NULL, colors=NULL, ncol=NULL, nrow=NULL, raster=FALSE, panel.size=NULL, adjust.func=NULL, title.size=6,raster.width=NULL, raster.height=NULL, adj.list=NULL, subset=NULL, ...) {
+plotEmbeddings <- function(embeddings, groups=NULL, colors=NULL, ncol=NULL, nrow=NULL, raster=FALSE, panel.size=NULL, adjust.func=NULL, title.size=6,raster.width=NULL, raster.height=NULL, adj.list=NULL, subset=NULL, return.plotlist=FALSE, ...) {
   if (is.null(panel.size)) {
     panel.size <- dev.size(units="in")
   } else if (length(panel.size) == 1) {
@@ -68,7 +72,7 @@ plotEmbeddings <- function(embeddings, groups=NULL, colors=NULL, ncol=NULL, nrow
   plot.list <- lapply(names(embeddings), function(n) {
     emb <- embeddings[[n]];
     if(!is.null(subset)) {
-      emb <- emb[rownames(emb) %in% subset,,drop=F]
+      emb <- emb[rownames(emb) %in% subset,,drop=FALSE]
     }
     embeddingPlot(emb, groups=groups, colors=colors, raster=raster,
                   raster.width=raster.width, raster.height=raster.height, ...) +
@@ -84,6 +88,8 @@ plotEmbeddings <- function(embeddings, groups=NULL, colors=NULL, ncol=NULL, nrow
   if (!is.null(adjust.func)) {
     plot.list <- lapply(plot.list, adjust.func)
   }
+  
+  if(return.plotlist) return(plot.list)
 
   return(cowplot::plot_grid(plotlist=plot.list, ncol=ncol, nrow=nrow))
 }
@@ -108,7 +114,7 @@ plotSamples <- function(samples, groups=NULL, colors=NULL, gene=NULL, embedding.
     }
   }
   if(class(embedding.type)=='matrix') { # actual embedding was passed
-    embeddings <- lapply(samples,function(r) embedding.type[rownames(embedding.type) %in% getCellNames(r),,drop=F])
+    embeddings <- lapply(samples,function(r) embedding.type[rownames(embedding.type) %in% getCellNames(r),,drop=FALSE])
     embeddings <- embeddings[unlist(lapply(embeddings,function(x) nrow(x)>0))]
   } else { # extract embeddings from samples
     embeddings <- lapply(samples, getEmbedding, embedding.type)
@@ -141,17 +147,20 @@ plotSamples <- function(samples, groups=NULL, colors=NULL, gene=NULL, embedding.
 #' @export
 plotClusterBarplots <- function(conos.obj=NULL, clustering=NULL, groups=NULL,sample.factor=NULL,show.entropy=TRUE,show.size=TRUE,show.composition=TRUE,legend.height=0.2) {
   ## param checking
-  if(!is.null(clustering)) {
-    if(is.null(conos.obj)) stop('conos.obj must be passed if clustering name is specified');
-    if(!clustering %in% names(conos.obj$clusters)) stop('specified clustering doesn\'t exist')
-    groups <- conos.obj$clusters[[clustering]]$groups
-  } else if (is.null(groups)) {
-    if(is.null(conos.obj)) stop('either groups factor on the cells or a conos object needs to be specified')
-    if(is.null(conos.obj$clusters[[1]])) stop('conos object lacks any clustering. run $findCommunities() first')
-    groups <- conos.obj$clusters[[1]]$groups
+  if (!is.null(clustering) && is.null(conos.obj)) {
+    stop('conos.obj must be passed if clustering name is specified')
   }
 
-  groups <- as.factor(groups)
+  if (is.null(groups) && is.null(conos.obj)) {
+    stop('Either groups factor on the cells or a conos object needs to be specified, both cannot be NULL')
+  }
+
+  groups <- parseCellGroups(conos.obj, clustering, groups)
+
+  if (!is.null(groups)) {
+    groups <- as.factor(groups)
+  }
+
   if(is.null(sample.factor)) {
     sample.factor <- conos.obj$getDatasetPerCell(); # assignment to samples
   }
@@ -172,8 +181,8 @@ plotClusterBarplots <- function(conos.obj=NULL, clustering=NULL, groups=NULL,sam
   pl <- list(clp + ggplot2::theme(legend.position="none"));
 
   if(show.entropy) {
-    if (!requireNamespace("entropy", quietly=T))
-      stop("You need to install 'entropy' package to use 'show.entropy=T'")
+    if (!requireNamespace("entropy", quietly=TRUE))
+      stop("You need to install 'entropy' package to use 'show.entropy=TRUE'")
 
     n.samples <- nrow(xt);
     ne <- 1-apply(xt, 2, entropy::KL.empirical, y2=rowSums(xt), unit=c('log2')) / log2(n.samples) # relative entropy
@@ -297,7 +306,7 @@ plotComponentVariance <- function(conos.obj, space='PCA',plot.theme=theme_bw()) 
 
   if(!is.null(pairs[[space]])) stop(paste("no pairs for space",space,"found. Please run buildGraph() or updatePairs() first, with score.component.variance=TRUE"))
 
-  nvs <- lapply(pairs,'[[','nv'); nvs <- setNames(unlist(nvs,recursive=F,use.names=F),unlist(lapply(nvs,names)))
+  nvs <- lapply(pairs,'[[','nv'); nvs <- setNames(unlist(nvs,recursive=FALSE,use.names=FALSE),unlist(lapply(nvs,names)))
   if(length(nvs)<1) stop("no variance information found. Please run buildGraph() or updatePairs() with score.component.variance=TRUE")
   if(space=='PCA') { # omit duplicates
     nvs <- nvs[unique(names(nvs))]
@@ -347,8 +356,8 @@ plotComponentVariance <- function(conos.obj, space='PCA',plot.theme=theme_bw()) 
 ##' @export
 plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,min.precision=NULL,n.genes.per.cluster=10,additional.genes=NULL,exclude.genes=NULL, labeled.gene.subset=NULL, expression.quantile=0.99,pal=colorRampPalette(c('dodgerblue1','grey95','indianred1'))(1024),ordering='-AUC',column.metadata=NULL,show.gene.clusters=TRUE, remove.duplicates=TRUE, column.metadata.colors=NULL, show.cluster.legend=TRUE, show_heatmap_legend=FALSE, border=TRUE, return.details=FALSE, row.label.font.size=10, order.clusters=FALSE, split=FALSE, split.gap=0, cell.order=NULL, averaging.window=0, ...) {
 
-  if (!requireNamespace("ComplexHeatmap", quietly = TRUE) || (substr(packageVersion("ComplexHeatmap"), 1, 1) <= "1")) {
-    stop("ComplexHeatmap >= 2.0 package needs to be installed to use plotDEheatmap. Please run \"devtools::install_github('jokergoo/ComplexHeatmap')\".")
+  if (!requireNamespace("ComplexHeatmap", quietly = TRUE) || packageVersion("ComplexHeatmap") < "2.4") {
+    stop("ComplexHeatmap >= 2.4 package needs to be installed to use plotDEheatmap. Please run \"devtools::install_github('jokergoo/ComplexHeatmap')\".")
   }
 
   groups <- as.factor(groups)
@@ -416,16 +425,16 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
       age <- do.call(rbind,lapply(sn(genes.to.add),function(gene) conos:::getGeneExpression(con,gene)))
       
       # for each gene, measure average correlation with genes of each cluster
-      acc <- do.call(rbind,lapply(expl,function(og) rowMeans(cor(t(age),t(og)),na.rm=T)))
-      acc <- acc[,apply(acc,2,function(x) any(is.finite(x))),drop=F]
+      acc <- do.call(rbind,lapply(expl,function(og) rowMeans(cor(t(age),t(og)),na.rm=TRUE)))
+      acc <- acc[,apply(acc,2,function(x) any(is.finite(x))),drop=FALSE]
       acc.best <- na.omit(apply(acc,2,which.max))
       
       for(i in 1:length(acc.best)) {
         gn <- names(acc.best)[i];
-        expl[[acc.best[i]]] <- rbind(expl[[acc.best[i]]],age[gn,,drop=F])
+        expl[[acc.best[i]]] <- rbind(expl[[acc.best[i]]],age[gn,,drop=FALSE])
       }
       if(additional.genes.only) { # leave only genes that were explictly specified
-        expl <- lapply(expl,function(d) d[rownames(d) %in% additional.genes,,drop=F])
+        expl <- lapply(expl,function(d) d[rownames(d) %in% additional.genes,,drop=FALSE])
         expl <- expl[unlist(lapply(expl,nrow))>0]
         
       }
@@ -435,7 +444,7 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
   # omit genes that should be excluded
   if(!is.null(exclude.genes)) {
     expl <- lapply(expl,function(x) {
-      x[!rownames(x) %in% exclude.genes,,drop=F]
+      x[!rownames(x) %in% exclude.genes,,drop=FALSE]
     })
   }
 
@@ -446,7 +455,7 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
 
   if(order.clusters) {
     # group clusters based on expression similarity (of the genes shown)
-    xc <- do.call(cbind,tapply(1:ncol(exp),groups[colnames(exp)],function(ii) rowMeans(exp[,ii,drop=F])))
+    xc <- do.call(cbind,tapply(1:ncol(exp),groups[colnames(exp)],function(ii) rowMeans(exp[,ii,drop=FALSE])))
     hc <- hclust(as.dist(2-cor(xc)),method='ward.D2')
     groups <- factor(groups,levels=hc$labels[hc$order])
     expl <- expl[levels(groups)]
@@ -459,7 +468,7 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
     # check if zoo is installed
     if(requireNamespace("zoo", quietly = TRUE)) {
       exp <- do.call(cbind,tapply(1:ncol(exp),as.factor(groups[colnames(exp)]),function(ii) {
-        xa <- t(zoo::rollapply(as.matrix(t(exp[,ii,drop=F])),averaging.window,mean,align='left',partial=T))
+        xa <- t(zoo::rollapply(as.matrix(t(exp[,ii,drop=FALSE])),averaging.window,mean,align='left',partial=TRUE))
         colnames(xa) <- colnames(exp)[ii]
         xa
       }))
@@ -494,7 +503,7 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
 
 
   if(!is.null(cell.order)) {
-    o <- cell.order[cell.order %in% colnames(o)]
+    o <- cell.order[cell.order %in% colnames(x)]
   } else { 
     o <- order(groups[colnames(x)])
   }
@@ -548,7 +557,15 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
     ra <- ComplexHeatmap::HeatmapAnnotation(df=rannot,which='row',show_annotation_name=FALSE, show_legend=FALSE, border=border,col=column.metadata.colors)
   } else { ra <- NULL }
 
-  #ComplexHeatmap::Heatmap(x, col=pal, cluster_rows=FALSE, cluster_columns=FALSE, show_column_names=FALSE, top_annotation=ha , left_annotation=ra, column_split=groups[colnames(x)], row_split=rannot[,1], row_gap = unit(0, "mm"), column_gap = unit(0, "mm"), border=T,  ...);
+  ## turns off ComplexHeatmap warning:
+  ## `use_raster` is automatically set to TRUE for a matrix with more than
+  ## 2000 columns. You can control `use_raster` argument by explicitly
+  ## setting TRUE/FALSE to it.
+  ## Set `ht_opt$message = FALSE` to turn off this message.
+  ## 
+  ht_opt$message = FALSE
+
+  #ComplexHeatmap::Heatmap(x, col=pal, cluster_rows=FALSE, cluster_columns=FALSE, show_column_names=FALSE, top_annotation=ha , left_annotation=ra, column_split=groups[colnames(x)], row_split=rannot[,1], row_gap = unit(0, "mm"), column_gap = unit(0, "mm"), border=TRUE,  ...);
   if(split) {
     ha <- ComplexHeatmap::Heatmap(x, name='expression', col=pal, cluster_rows=FALSE, cluster_columns=FALSE, show_row_names=is.null(labeled.gene.subset), show_column_names=FALSE, top_annotation=ha , left_annotation=ra, border=border,  show_heatmap_legend=show_heatmap_legend, row_names_gp = grid::gpar(fontsize = row.label.font.size), column_split=groups[colnames(x)], row_split=rannot[,1], row_gap = unit(split.gap, "mm"), column_gap = unit(split.gap, "mm"), ...);
   } else {
@@ -571,3 +588,4 @@ plotDEheatmap <- function(con,groups,de=NULL,min.auc=NULL,min.specificity=NULL,m
 
   return(ha)
 }
+