+++ CHANGELOG.md
@@ -9,6 +9,16 @@
 
 ### Added
 
+- Changed `std::cout` to `Rcpp::Rcout` (July 2020)
+- Revised README, vignettes (July 2020)
+- added `ht_opt$message = FALSE` for ComplexHeatmap (July 2020)
+- Write to HDF5 for `saveConosForScanPy()` (7 July 2020)
+- LICENSE (July 2020)
+
+### Removed
+
+- Removed `getCorrectionVector()` and `getPerCellTypeDECorrected` (2 July 2020)
+- Removed all neighborhood averaging via `neighborhood.average` (4 July 2020)
 
 ## [1.3.0] - 2020-19-3